refactor: migrate effect module
